requirements.txt
psycopg2-binary
sql-metadata>=2.7.0
SQLAlchemy==1.4.54
sqlparse==0.5.0
Unidecode==1.3.8
pyyaml==6.0.2
discord==2.3.2
ratelimit==2.2.1
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.4
pytz==2025.2
tzdata==2025.2
requests==2.31.0
jinja2>=3.1.3,<4
python-dotenv==1.0.0
protobuf>=5,<6
pytest==7.4.3
pytest-mock==3.12.0